New network configuration
Update instructions and scripts to use new network configuration where
InfluxDb and Grafana talk to each other in their own separate network.
This allows configuring connection between these containers using
container names as hostnames instead of ip addresses which is better
because ip addresses can change depending on the order in which the
containers are started for example on reboot.

Similar isolation to their own network for RuuviCollector and InfluxDb
is not possible due to RuuviCollector requiring host network access to
listen on bluetooth devices.