Backup and restore influxdb during maintenance
Change how influxdb is maintained when updating container. Previously
data was stored in docker volume and the container was updated
separately and the new instance started to use the old data. Backups of
the data weren't actually instructed at all (and were never made).

This update makes changes to scripts and instructions to change the
maintenance flow to such that first a backup of the data is made, then
new version of influxdb is started without any data and data is restored
from the backup. This provides two benefits:
1. no risk of a bug in migration breaking something
2. backup and restoration gets properly tested on every maintenance run